% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{fixed_value_simulation}
\alias{fixed_value_simulation}
\title{Simulate fixed values for a given number of clusters}
\usage{
fixed_value_simulation(C, ...)
}
\arguments{
\item{C}{An integer. The number of clusters.}

\item{...}{Additional arguments if need to change distributions of the fixed terms}
}
\value{
A list containing the following elements:
\itemize{
\item mu_c: A numeric vector. The overall cluster mean. \in R^{C x 1}
\item n_c: A numeric vector. The number of observations in each cluster. \in R^{C x 1}
\item n: An integer. The total number of observations. \in R
\item W_p: A numeric vector. The probability of being untreated. \in R^{C x 1}
\item W_c: A numeric vector. The cluster level treatment assignment. W_c = 1 => never treated cluster \in R^{C x 1}
\item indi_term: A numeric vector. The individual term. \in R^{n x 1}
\item cluster: A numeric vector. The cluster. \in R^{n x 1}
\item never_trt: A numeric vector. The never treated. \in R^{n x 1}
\item trt: A numeric vector. The treatment. \in R^{n x 1}
}
}
\description{
This function simulates fixed values for a given number of clusters (C). It generates
the overall cluster mean, number of observations in each cluster, total number of observations,
probability of being untreated on the cluster level, cluster level treatment assignment, individual term, cluster,
never treated, and treatment.
}
\examples{
\dontrun{
fixed_value_simulation(5)
}
}
