% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{helper_indi_term_c}
\alias{helper_indi_term_c}
\title{Generate individual term based on cluster term and normal distribution}
\usage{
helper_indi_term_c(c, mu_c, n_c)
}
\arguments{
\item{c}{An index for selecting the cluster term from \code{mu_c} and the number of terms from \code{n_c}.}

\item{mu_c}{A vector of cluster terms.}

\item{n_c}{A vector of the number of individual terms to generate for each cluster.}
}
\value{
A vector of individual terms.
}
\description{
This function generates an individual term based on a cluster term and a normal distribution.
The cluster term is selected from \code{mu_c} based on the index \code{c}, and the individual term is generated
by drawing from a normal distribution with mean equal to the cluster term and standard deviation 5.
The number of individual terms generated is determined by \code{n_c[c]}.
}
