% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{potential_outcome_df}
\alias{potential_outcome_df}
\title{Generate a data frame of potential outcomes over time}
\usage{
potential_outcome_df(
  t_total,
  fixed_res,
  treatment_df,
  cluster_coef,
  indi_coef,
  past_coef,
  long = TRUE
)
}
\arguments{
\item{t_total}{The total time period (excluding t = 0).}

\item{fixed_res}{A list containing fixed results from the 'fixed_value_simulation' function.}

\item{treatment_df}{A data frame containing the treatment info G, cluster, G_min, and G_mod.}

\item{cluster_coef}{The coefficient for the cluster effect.}

\item{indi_coef}{The coefficient for the individual effect.}

\item{past_coef}{The coefficient for the past effect.}

\item{long}{Logical indicating whether to return the data in long format. If TRUE, the data is returned
in long format with one row per time point per individual. If FALSE, the data is returned in wide format
with one row per individual and one column per time point. Default is TRUE.}
}
\value{
A data frame of potential outcomes over time. If \code{long} is TRUE, the data frame also includes
the treatment effect and Y_t_inf at each time point.
}
\description{
This function generates a data frame of potential outcomes over time based on the treatment effect,
the simulated Y_t_inf, and the treatment data. The potential outcomes are calculated as the sum of
Y_t_inf and the treatment effect. The function can return the data in either long or wide format.
}
\examples{
t_total <- 10
C = 10 
fixed_res <- fixed_value_simulation(C)
treatment_df <- treatment_time_G(t_total, fixed_res)
cluster_coef <- 3
indi_coef <- 3
past_coef <- 0.5
potential_outcome_df(t_total, fixed_res, treatment_df, cluster_coef, indi_coef, past_coef)
}
