% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_sim}
\alias{run_sim}
\title{Run many simulations in parallel}
\usage{
run_sim(n_sims, n_cores = 1, ...)
}
\arguments{
\item{n_sims}{Number of simulations to run}

\item{n_cores}{Number of cores}

\item{...}{Simulation arguments}
}
\description{
Run many simulations in parallel
}
