% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_Y_t_inf}
\alias{sim_Y_t_inf}
\title{This function simulates a dataframe of potential outcomes under control for a given number of time points such that Y_t(inf) = indi_term + v_t, where v_0 = norm(n, 0,1) and v_t = norm(n, v_t_minus_1, 1) for t = 1, 2, ..., t_total.}
\usage{
sim_Y_t_inf(t_total, fixed_res, beta = 0)
}
\arguments{
\item{t_total}{The total number of time points (excluding t = 0) for the simulation}

\item{fixed_res}{A list returned from the 'fixed_value_simulation' function.
The list 'res' should contain the following elements: "n", "indi_term".}
}
\value{
A data frame where each column represents the Y_inf at a different time point \in R^{n x t_total + 1}]}
}
\description{
This function simulates a dataframe of potential outcomes under control for a given number of time points such that Y_t(inf) = indi_term + v_t, where v_0 = norm(n, 0,1) and v_t = norm(n, v_t_minus_1, 1) for t = 1, 2, ..., t_total.
}
