% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{treatment_effect}
\alias{treatment_effect}
\title{Calculate the treatment effect over time}
\usage{
treatment_effect(
  t_total,
  fixed_res,
  treatment_df,
  cluster_coef,
  indi_coef,
  past_coef
)
}
\arguments{
\item{t_total}{The total time period (excluding time = 0).}

\item{fixed_res}{A list containing fixed results from the 'fixed_value_simulation' function.}

\item{treatment_df}{A data frame containing the treatment info G, cluster, G_min, and G_mod.}

\item{cluster_coef}{The coefficient for the cluster effect.}

\item{indi_coef}{The coefficient for the individual effect.}

\item{past_coef}{The coefficient for the past effect.}
}
\value{
A data frame containing the treatment effect at each time point.
}
\description{
This function calculates the treatment effect over time based on the proportion of treatment,
individual treatment, and past treatment effect. The treatment effect at each time point is
calculated as the sum of the cluster effect, individual effect, and past effect.
}
