% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{treatment_time_G}
\alias{treatment_time_G}
\title{Simulate treatment distribution}
\usage{
treatment_time_G(t_total, fixed_res, ..., treat_prob_fun = NULL)
}
\arguments{
\item{t_total}{The total number of time points (excluding t = 0) for the simulation}

\item{fixed_res}{A list containing fixed results from the 'fixed_value_simulation' function.}

\item{...}{Additional arguments passed to the function.}

\item{treat_prob_fun}{A function that takes 't_total' as input and returns a vector of treatment probabilities.
If NULL (the default), all time points have equal probability of treatment.}
}
\value{
A data frame with n rows and columns 'G' (the treatment time), 'cluster' (the cluster ID), 'Gmin' (the minimum treatment time in the cluster), and 'G_mod' (the modified treatment time, with 't_total + 1' indicating never treated).
}
\description{
This function simulates the distribution of treatment times for a set of units.
By default, all units have the same probability of being treated at any time point     #' from 1:t_total,
but a custom treatment probability function can be provided.
}
